Crate ethcore_ipc_nano

IPC over nanomsg transport

Structs

DEBUG_FLAG
Set PARITY_IPC_DEBUG=1 for fail-fast connectivity problems diagnostic

GuardedSocket
struct for guarding _endpoint (so that it wont drop) derefs to client S

NanoSocket
A type-safe socket wrapper around nanomsg's own socket implementation. This provides a safe interface for dealing with initializing the sockets, sending and receiving messages.

Worker
Generic worker to handle service (binded) sockets

Enums

SocketError
Error occurred while establising socket or endpoint

Traits

IpcConfig
Allows to configure custom version and custom handshake response for ipc host

IpcInterface
Allows implementor to be attached to generic worker and dispatch rpc requests over IPC

WithSocket
Basically something that needs only socket to be spawned

Functions

client
Spawns client <S> over specified address creates socket and connects endpoint to it for request-reply connections to the service

fast_client
Client over interface that is supposed to give quick almost non-blocking responses

generic_client
Client with no default timeout on operations

init_duplex_client
Spawns client <S> over specified address creates socket and connects endpoint to it for duplex (paired) connections with the service